Major Update
A lot happened between the last update. got connection.php made to make a connection from our site to our database Got functions.php made for our function to check for a user session then return their data under our member table to use to refer later to other tables in our query. We have index.php made for our home page, which is different for members with accounts and guests, with members being able to view what cars they have saved, as well as be able to go to their account settings. Account.php was made that will be worked on to let the user be able to change their settings/contant info. Index.php still in progress with integrating all features our database has to offer, as well as setting up our part search system and our system for the users to be able to add their own custom vehicles to the database as well.