Major Update
Made an account.php for users to be able to update their password and view their saved info. Added the ability for users to remove vehicles from their inventory, as well as add to their inventory from vehicles we have in stock. Working on database this site uses so I can set up the search queries for parts and their corresponding vehicles. Also need to add links to the about and contact page, as well as the page themselves